FEAT(pkt): use wait time from config file as msec

+++ cmd/anyclient/etc/etc.go
@@ -24,7 +24,7 @@ type AnyClientConfig struct {
 	} `yaml: "Service"`
 	Packet struct {
 		Filename string `yaml: "filename"`
-		Wait     uint8  `yaml:"wait"`
+		Wait     uint32 `yaml:"wait"`
 	} `yaml: "packet"`
 	Archive struct {
 		Filename string `yaml: "filename"`
